source_code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
gettext=0.19.8.1=h1a89ca6_1005
gtts=2.2.3=pypi_0
image=1.5.33=pypi_0
jpeg=9d=h8ffe710_0
openjpeg=2.4.0=hb211442_1
pdf2image=1.15.1=pypi_0
pillow=8.2.0=pypi_0
playsound=1.2.2=pypi_0
poppler=21.03.0=h9ff6ed8_0
poppler-data=0.4.10=0
pygame=2.0.1=pypi_0
pygments=2.9.0=pyhd3eb1b0_0
pytesseract=0.3.7=pypi_0
tesseract=0.1.3=pypi_0
tk=8.6.10=h8ffe710_1
tkhtmlview=0.1.0.post1=pypi_0
tkinterweb=3.9.11=pypi_0
